```
subweight compare commits \
>          --path-pattern "./system-parachains/bridge-hubs/**/weights/**/*.rs" \
>          --format markdown --no-color  \
>          --change added changed removed  \
>         --method asymptotic --ignore-errors  \
>         --strip-path-prefix="system-parachains/bridge-hubs/" \
>         --threshold 0.5 \
>          remotes/Snowfork/ethereum-headers-on-demand \
>          origin/bko-snowbridge-benchmarks
```

| File                                                                 | Extrinsic                  | Old      | New      | Change [%] |
|----------------------------------------------------------------------|----------------------------|----------|----------|------------|
| bridge-hub-polkadot/src/weights/snowbridge_pallet_ethereum_client.rs | submit                     | 25.40ms  | 34.32ms  | +35.11     |
| bridge-hub-polkadot/src/weights/snowbridge_pallet_ethereum_client.rs | submit_with_sync_committee | 127.21ms | 167.58ms | +31.73     |
| bridge-hub-polkadot/src/weights/snowbridge_pallet_ethereum_client.rs | force_checkpoint           | 102.44ms | 133.80ms | +30.61     |
| bridge-hub-kusama/src/weights/snowbridge_pallet_ethereum_client.rs   | submit                     | 26.22ms  | 33.01ms  | +25.89     |
| bridge-hub-kusama/src/weights/snowbridge_pallet_ethereum_client.rs   | submit_with_sync_committee | 131.31ms | 162.24ms | +23.56     |
| bridge-hub-kusama/src/weights/snowbridge_pallet_ethereum_client.rs   | force_checkpoint           | 105.80ms | 129.70ms | +22.58     |
| bridge-hub-kusama/src/weights/snowbridge_pallet_inbound_queue.rs     | submit                     | 481.28us | 584.43us | +21.43     |
| bridge-hub-polkadot/src/weights/snowbridge_pallet_inbound_queue.rs   | submit                     | 482.86us | 585.13us | +21.18     |