@claravanstaden 2024-07-02
Thanks for the fresh benchmarks @bkontur! I am concerned about the increase in weight, especially for all the ethereum client extrinsics. They should start more or less the same.

The diff shows that there were mostly minor removals in `[bridges/snowbridge/parachain/pallets/ethereum-client/src/lib.rs](https://github.com/paritytech/polkadot-sdk/compare/e046d3b2caaca6c43a60f26594052c5c806b980e..6dba07c0bc0c14c2caa33dc8fc385c849175f6d5#diff-98c747bd18857fe76701570ede9bc3bef5350ed0f9808daea2af1e769c1a8eff)`. Have you seen weight increased like this before, for code that didn't really change?