@acatangiu 2024-07-02
@claravanstaden can you run short benchmark for just one of them with old version and new, on your local machine and see if you are getting any change between them?

if so, someone on your team can also "debug" the benchmarking code for said extrinsics and see if it is actual real-world cost change or benchmarking issue