@claravanstaden 2024-07-03
@bkontur @acatangiu thanks for running the benchmarks again! It looks like the diff might be from moving the hash tree root of the finalized header: https://github.com/paritytech/polkadot-sdk/compare/e046d3b2caaca6c43a60f26594052c5c806b980e..6dba07c0bc0c14c2caa33dc8fc385c849175f6d5#diff-98c747bd18857fe76701570ede9bc3bef5350ed0f9808daea2af1e769c1a8effR502

I'm going to merge the PR.